+++ webextension/js/classes/chrome-native-settings.js
@@ -1,0 +1,4 @@
+export const chromeNativeSettingsStorageKey = '_nativeSettings';
+export async function getElectronSettings() {
+	return (await browser.storage.local.get([chromeNativeSettingsStorageKey])) ?? {};
+}